OeX_ES-33: Renames edx_haystack_extensions application name on a new one edx_elasticsearch_dsl_extensions.
