+++ course_discovery/apps/course_metadata/search_indexes/serializers/person.py
@@ -2,7 +2,7 @@ from django_elasticsearch_dsl_drf.serializers import DocumentSerializer
 from rest_framework import serializers
 
 from course_discovery.apps.api.serializers import ContentTypeSerializer, MinimalPersonSerializer
-from course_discovery.apps.edx_haystack_extensions.serializers import BaseDjangoESDSLFacetSerializer
+from course_discovery.apps.edx_elasticsearch_dsl_extensions.serializers import BaseDjangoESDSLFacetSerializer
 from .common import DocumentDSLSerializerMixin
 from ..constants import BASE_SEARCH_INDEX_FIELDS, COMMON_IGNORED_FIELDS
 from ..documents import PersonDocument
