+++ course_discovery/apps/edx_catalog_extensions/api/serializers.py
@@ -4,7 +4,7 @@ from rest_framework.fields import DictField
 
 from course_discovery.apps.api.serializers import QueryFacetFieldSerializer
 from course_discovery.apps.course_metadata.search_indexes.serializers.aggregation import AggregateFacetSearchSerializer
-from course_discovery.apps.edx_haystack_extensions.serializers import (
+from course_discovery.apps.edx_elasticsearch_dsl_extensions.serializers import (
     FacetDictField,
     FacetFieldSerializer,
     FacetListField,
