+++ course_discovery/apps/edx_elasticsearch_dsl_extensions/elasticsearch_boost_config.py
