+++ course_discovery/apps/edx_elasticsearch_dsl_extensions/management/commands/remove_unused_indexes.py
