+++ course_discovery/apps/edx_elasticsearch_dsl_extensions/management/commands/update_index.py
