+++ course_discovery/apps/edx_elasticsearch_dsl_extensions/serializers.py
