+++ course_discovery/apps/edx_elasticsearch_dsl_extensions/tests/test_remove_unused_indexes.py
